+++ src/scriptlet.h
@@ -1,8 +1,6 @@
 #ifndef SCRIPTLET_H
 #define SCRIPTLET_H
 
-#include <stdint.h>
-
 typedef struct _scriptlet* scriptlet;
 typedef struct _runtime* runtime;
 
@@ -14,10 +12,10 @@ extern scriptlet scriptlet_open(const char* text);
 
 extern runtime scriptlet_prepare(scriptlet s);
 
-extern int scriptlet_set_int(runtime r, const char* k, int64_t v);
+extern int scriptlet_set_int(runtime r, const char* k, long long v);
 extern int scriptlet_set_real(runtime r, const char* k, double v);
 extern int scriptlet_set_string(runtime r, const char* k, const char* v);
-extern int scriptlet_get_int(runtime r, const char* k, int64_t* v);
+extern int scriptlet_get_int(runtime r, const char* k, long long* v);
 extern int scriptlet_get_real(runtime r, const char* k, double* v);
 extern int scriptlet_get_string(runtime r, const char* k, const char** v);
