+++ src/tree.h
@@ -2,7 +2,6 @@
 #define TREE_H
 
 #include <stdlib.h>
-#include <stdint.h>
 
 #ifndef TREE_KEY_USERDEFINED
 #include "uuid.h"
@@ -14,14 +13,14 @@ typedef struct _tree* tree;
 
 extern tree tree_create(void);
 
-extern int tree_add(tree tptr, const TREE_KEY* key, uint64_t value);
-extern int tree_get(const tree tptr, const TREE_KEY* key, uint64_t* value);
-extern int tree_set(const tree tptr, const TREE_KEY* key, uint64_t value);
+extern int tree_add(tree tptr, const TREE_KEY* key, unsigned long long value);
+extern int tree_get(const tree tptr, const TREE_KEY* key, unsigned long long* value);
+extern int tree_set(const tree tptr, const TREE_KEY* key, unsigned long long value);
 extern int tree_del(tree tptr, const TREE_KEY* key);
 
 extern size_t tree_count(const tree tptr);
 extern int tree_stats(const tree tptr, long* trunks, long* branches, long* leafs);
-extern size_t tree_iter(const tree tptr, void* h, int (*f)(void* h, const TREE_KEY* key, uint64_t* value));
+extern size_t tree_iter(const tree tptr, void* h, int (*f)(void* h, const TREE_KEY* key, unsigned long long* value));
 
 extern void tree_destroy(tree tptr);
